feat: реализация семафора через пакет /x/sync

+++ internal/channel/Semaphore.go
@@ -4,7 +4,7 @@ type Semaphore struct {
 	weight chan struct{}
 }
 
-func NewSemaphore(maxWeight *int32) *Semaphore {
+func NewSemaphore(maxWeight *int64) *Semaphore {
 	sem := &Semaphore{
 		weight: make(chan struct{}, *maxWeight),
 	}
